feat: move rest of theme component
