+++ packages/strapi-design-system/src/themes/ColorButton.tsx
@@ -1,14 +1,17 @@
-import React from 'react';
-import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { Box } from '../Box';
+
+import { Box, BoxProps } from '../Box';
 import { Typography } from '../Typography';
 
 const StyledBox = styled(Box)`
   background: ${({ color }) => color};
 `;
 
-export const ColorButton = ({ colorKey, color }) => {
+interface ColorButtonProps extends Pick<BoxProps, 'color'> {
+  colorKey: string;
+}
+
+export const ColorButton = ({ colorKey, color }: ColorButtonProps) => {
   return (
     <StyledBox color={color} paddingTop={6} paddingLeft={6} hasRadius shadow="popupShadow" borderColor="neutral200">
       <Box background="neutral0" padding={1}>
@@ -16,9 +19,4 @@ export const ColorButton = ({ colorKey, color }) => {
       </Box>
     </StyledBox>
   );
-};
-
-ColorButton.propTypes = {
-  color: PropTypes.string.isRequired,
-  colorKey: PropTypes.string.isRequired,
 };
