+++ packages/strapi-design-system/src/types/utils.ts
@@ -1,0 +1,3 @@
+export type PickType<TObject, TKey extends keyof TObject> = {
+  [P in TKey]: TObject[P];
+}[TKey];
