+++ packages/strapi-icons/tsconfig.json
@@ -5,7 +5,8 @@
     "moduleResolution": "Node",
     "useDefineForClassFields": true,
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "allowJs": false,
+    // change back to false when complete
+    "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": false,
     "allowSyntheticDefaultImports": true,
